+++ src/Utility.cpp
@@ -316,4 +316,11 @@ std::shared_ptr<Vector2d> lineLineIntersect2d(
 
   return intersect_pt;
 }
+
+Vector3d slerp(const Vector3d &from, const Vector3d &to, double s){
+  double theta = acos(from.dot(to));
+  double alpha = sin((1-s)*theta)/sin(theta);
+  double beta = sin(s*theta)/sin(theta);
+  return alpha*from+beta*to;
+}
 } // namespace cg
